+++ matchmaker-lib/src/lib.rs
@@ -26,8 +26,9 @@ pub mod tui;
 // misc
 mod aliases;
 pub mod errors;
-mod utils;
+pub mod utils;
 pub use aliases::*;
 pub use errors::*;
+pub use ratatui::text::{Line, Span, Text};
 
 pub mod noninteractive;