+++ matchmaker-lib/src/utils/mod.rs
@@ -1,7 +1,9 @@
 mod percentage;
 pub use percentage::Percentage;
 
+pub mod crossterm;
 pub mod serde;
+pub mod string;
 pub mod text;
 
-pub mod string;
+pub use crossterm::{color_to_crossterm, span_to_ansi, style_to_crossterm, text_to_ansi};